Improve building selection menu
Instead of displaying an arrow on the bottom of the screen when the
menu icons overflow, now icons can overflow from both ends (top and
bottom), regardless of the position of the status bar (top or bottom)
by disabling the sprite display.

This means that when the building selection menu is displayed sprites
can't be shown on the status bar.