Prevent reads from ECHO RAM when scrolling the BG
The BG scrolling functions don't check the bounds of the source
buffer. This means that when it was trying to scroll to the top of
the map or the bottom of it the reads would overflow from the top or
the bottom of the buffer. In the case of a map loaded to WRAMX, this
means overflowing from DFFFh into ECHO RAM, causing emulators such as
BGB to complain about it.