Fix flicker when hiding building selection menu
When the status bar was located on the bottom of the screen, the same
frame the building selection menu was hidden, the sprites supposedly
covered by the status bar reappeared for one frame. Note that since
the status bar is displayed using the window, and it doesn't hide the
sprites, this has to be done by software.