Added hourglass icon, mirrored arrow sprite
Added hourglass icon to tell the user when the CPU is still simulating
a step in situations when it is needed to end to do something. For
example, when trying to build something, if the simulation step hasn't
finished, the hourglass will be shown until it has finished. When
entering the pause menu it will also be shown (minimap room cannot be
entered if a simulation is still running).

The arrow icon used in the building selection menu now uses 2 tiles
instead of 4 to save space, as the hourglass uses 2 tiles as well.
The arrow sprites are now mirrored to form the two halves.