Fix animation not happening sometimes
Previously, when moving the cursor around the screen the simulation
would hang. Now, it only stops when the bg is actually being
scrolled, which is the only real case in which the animation
shouldn't be updated.

BG conditional scrolling functions have been modified to return
whether they have scrolled or not.