Fix CPU busy icon not showing/hiding
Added a safeguard to CPUBusyIconHandle to hide the icon when needed,
and made the code in room_game.asm to show the icon always. This way
it is the responsibility of CPUBusyIconHandle to hide the icon when
the simulation ends, while room_game.asm can still hide the icon
manually.