Implement generic save system
Save integrity checks are not implemented yet and not all the
information is saved yet, but this system is already scalable to an
undefined number of SRAM banks.

The information about the save data is in `save_struct.inc`. Each
field has a name and a size, and they are all copied from WRAM with
no manipulation except for the attribute map. Only the BANK1 bit is
saved in this case, and 8 bits are packed (for 8 tiles) and stored
in order to save memory. The palette is recalculated at load time.