Fix string not accessible from ROMX
STR_CITY_NAME is stored in ROMX, and it was used from ROMX in a
different section. Sometimes they were assembled into the same bank,
but not always. This caused garbage to be printed instead of the
string. Now, the string is saved into the stack before calling the
other function so that it can read it.