Prevent building bridges that wrap around the map
Instead of checking if the final coordinates are inside the map, check
each step. The function that returns the type of the map returns water
or field even if the coordinates are outside the map. If both sides
have water, it could wrap around the map and enter from the other
side. Since the check was only done at the end, and the last tile is
inside the map, this would be considered a valid bridge.

This only happened when building horizontal bridges.