Power line bridges can be burned
They don't make the fire go to the other side of the bridge, they
just disappear.

Bridge delete function modified to accept being called with or without
money check. Also, the VRAM map is not updated from the bridge delete
function because it isn't thread safe and this can cause VRAM
corruption.