Finished historical population graph
The autoscale works like this: The scale represents the amount of bits
shifted to scale the population down to the graph Y axis. First, the
code tries to fit the current population in 127. If it doesn't fit, it
increases the scale and shifts all the stored values by 1 (divide by
2), checking if it fits then. If so, the new value is stored.

The current scale is also saved along with the array of entries to
know how to scale new values added to the graph.

This way, the graph represents more valuable information than having
a fixed scale, while not too much information is stored in the SRAM
(only 1 byte per entry).