Replace methods for including strings by charmaps
Instead of having a few different systems to convert from ASCII
strings in the source code to tile indexes in the final ROM (or in
VRAM in runtime) use the charmap to preconvert everything when
assembling the ROM.

Removed ASCII2Tile and the array it used to convert from ASCII to
tile index. This way, some space is saved in ROM0.