Pause menu rooms return to menu instead of map
When returning from any of the rooms inside the ingame pause menu,
instead of returning to the city map, the game returns to the pause
menu. This way, instead of forcing the player to wait for another
simulation step to end, the user can do anything else in this menu
right away.

Some changes to the loading code have been made to allow this
behaviour.