Fix status bar transitions
When closing the pause menu, the status bar is not shown for a frame,
causing a visual glitch that can be annoying.

Similarly, when there are more than 1 messages in the queue, when
closing the window of the first one, there is a frame in which the
window is closed and the status bar is shown right before the box
pops up again. Now, when trying to close the box, if there are more
messages in the queue, instead of closing the box the new message is
writen instead.