Refresh BG during fires correctly
Previously, during a fire, when opening the build menu and trying to
build, it could happen that the fire wasn't updated and the first try
would fail because there is fire in the location the user is trying
to build.

This happened because the BG update was left to the animation handler,
which stops refreshing the BG when in build mode.

With this patch, the BG is always updated after simulating a step,
so that the BG always shows the actual state of the map.