Implement city technology level
Cities now have a hidden counter of technology level that enables
certain buildings (nuclear and fusion power plants for now).

This counter tries to increase at the end of each year. For each
university in the map the counter has the possibility of being
incremented by one.

Added text messages to inform the user about new discoveries and
insufficient technology to build buildings.

The value is saved to SRAM to preserve it.