Recalculate status variables before main loop
There are some things that aren't saved with the rest of the city data
in the SRAM. Before, the game had to go through a full simulation step
before refreshing them.

With this patch, some things, like the total population of the city,
the city title and the RCI demand are updated before the first
iteration. Also, this prepares the code for refreshing the number of
buildings of each type for future optimizations.