Small fix to VBL handler in room game
There's a flag that prevents most of the code in the VBL handler from
being executed twice at the same time. This patch makes the flag be
set right after testing if it is set or not. This isn't a real
problem right now because there's more than enough time to set the
flag after the test, but this way the code is easier to understand.