Fix return values of cursor/map moving functions
The cursor/map moving functions and scrolling functions are only
supposed to return 1 after a scroll step where a new row/column is
loaded.

Previously the functions returned 1 in some cases when there were no
load of a new line. This unwanted behaviour only happened when the
user tried to scroll past the limit of the map.

This should also fix the really weird case in which the user may
scroll past the end of the map and break the scroll system.