Initial implementation of transportation objects
When in the city map, in watch mode or fast move mode, things like
planes, trains and boats may spawn and move around for decoration.

This commit implements the base system for all of them and the code
for planes, as well as some placeholders for trains. It also adds the
calls to this module to use it.

This is a work in progress, and there are still a few left to do,
like updating the number of enabled planes after building or
demolishing airports.