Don't switch game states during scroll
Make the game wait until the scroll is 8-pixel aligned before changing
between watch and fast movement modes. If the user holds or releases
the B button the game will perform a bg scroll drift until both scroll
coordinates are aligned and then allow the game to switch to the other
mode.