Make status bar hide transportation sprites
The variable that specifies if sprites are shown or not over the
status bar has been moved to the file with the status bar code instead
of the one that contains the menu to select buildings.

Instead of changing the value only when entering or exiting the
building selection menu, the value is changed whenever the game state
changes, which makes it easier to control the result.

This way, planes and other transportation sprites aren't shown over
the status bar. Sprites are only allowed over it when in edit mode,
and during that mode the transportation animations are disabled. The
only sprites visible during edit mode are the cursor and the icon of
the building being built.