Fix irregular transport animations
Due to the way animations are handled, it was impossible for the game
to keep up with the speed that the animations of transportation
objects required. This resulted in a noticeable pause in the animation
of the planes once every second, when the rest of the map (traffic,
fire, etc) is handled. Because of the limitations of the old system,
a new one has been implemented.

Instead of handling animations during the VBL handler (after the
critical section at the beginning) the handling function has been
split into two functions.

The first one is the VBL handler, which does simple movement handling
and doesn't actually destroy or create objects, which takes time. The
second one is the game step handler, that checks if an object should
be destroyed and moved to another place, etc. This creation and
destruction of objects takes time, so it is better to handle it
sepparately from the critical parts of the animation handling.