Implement correct train movement routines
Trains move around the train tracks, change direction sometimes (and
at the corners). Trains disappear when they leave the track for any
reason, which is useful in the way that the train sprites are shown
in different parts of the city in case there are different train
circuits but the trains are generated in only one of them.