Add a margin to the cursor scroll
Instead of scrolling when the cursor reaches the border of the screen,
scroll when there are 4 tiles in between, only reach the border of the
screen when the background can't be scrolled more.

If the cursor is in that extra border and the background is scrolled
back (holding the B button so that the cursor doesn't move), the
cursor will be stuck in that position even if it shouldn't be there,
but it won't advance anymore. Of course, it will work normally in the
other direction.