Optimize technology advancements simulation
Precalculate number of universities so that the code doesn't have to
search the map for any of them every time the code is run.

This is only more optimal if the user just lets the simulation run
undisturbed, if the user builds something, there's a fire, etc, the
number of buildings is recalculated so it can actually be slower after
a few changes like the ones said.

The difference is that the number of buildings is calculated at times
when the user has to wait for other reasons, so this performance
reduction (which is really small in comparison with the old code) in
that case won't be noticeable. It will make the step from December to
January a bit faster, though, so it is worth the change.