Fix boats not spawning in some cases
Previously, if a boat couldn't spawn in the first try, it wouldn't
spawn at all.

Now, when the boat handling system is reseted, a new array is filled
to tell the boats that are enabled or disabled, and that array is
checked every frame to spawn boats that are enabled but not spawned.