Add enable/disable option for animations and music
The options menu has been updated to show the new options, even though
they are not functional right now (music isn't even implemented).

Animations need a lot of CPU power so, if the player decides they are
not necessary, they can be disabled so that the simulation can take
advantage of it and make the in-game time run faster.

Similarly, music may not be necessary if the player is listening to
music from other device, so it can be disabled to increase the speed
of the simulation (when music is actually implemented!).