When measuring SRAM size use its last byte to test
The code used to check the number of available banks of SRAM used to
save the value of to the first byte of each SRAM bank, overwrite it,
and then restore it.

Doing it may corrupt other game's save data in flashcarts if this game
is run by mistake, and somehow the GB is turned off while checking.

This change makes it less likely to corrupt actual save data: If
there's unused space in a bank, it's usually at the end.