Fix initial spawn of transportation objects
To fix this it is needed to split the main bg loading function into
one that loads the map and tile data and another one that loads the
palettes. If palettes are loaded right after loading the map data
there will be an instant during which only the bg is visible, and the
transportation sprites pop up right afterwards. This is not the
desired behaviour, everything should appear at the same time.

Now, the bg data is loaded, then the sprite palettes are loaded, and
then the shadow OAM is updated with the transportation objects. The
actual OAM isn't updated until the VBL handler is called, which will
only happen after the bg palettes are loaded. The difference of time
between those two is really small, so the popping up problem doesn't
happen.