Use channel 1 for SFX instead of channel 2
Channel 1 has sweep support, which is useful for SFX. The code to use
channel 2 has been disabled because it is easier to make songs if the
channels used for SFX are always the same. This way, the main melody
can be placed in a channel that is always unused by SFX so that it
isn't interrupted when a SFX is triggered.