Implement game over code
After 4 quarters with negative budget and negative funds game over is
reached. After the game over, the simulation pauses forever but it
doesn't exit to the main menu.

The count of quarters is saved to SRAM.

Also, split message of running out of money into two depending on
whether the player has a pending loan or not.