Add population checks for building construction
Now, buildings can be restricted to specific city sizes. If a building
is not allowed, an error message is shown when trying to build it.

Stadium, port and airport have been restricted to city class or higher.
It was problematic before because they generated money on their own so
a city could be composed of this kind of building only and be viable,
even from the start.