Fix train spawn coordinates generation
The search loop decreases the value in DE by one whenever the loop finds
a train track tile in the map. When the counter is 0, the next train
tile that is found is the spawn location of the train.

Previously, the loop decremented DE before checking if it was 0, which
resulted in it being considered FFFFh instead and triggering a debug
breakpoint.

The bug didn't cause any problem because the only consequence is that
the train didn't spawn during that frame and it tried to respawn again
the following one.