Simplify map generation room controls
Use pad buttons UP and DOWN instead of using SELECT to switch between
'Land' and 'Water'. This is more intuitive, but now only RIGHT and LEFT
can be used to modify the seed, which is a bit slower because they only
increase in 1 increments and there's no way to have increments of 10.

The background image has been simplified and the GUI refresh code has
been update to the new coordinates of the map type cursor and seed
number.