Prevent closing room game messages by mistake
Previously, if a message apeared but the user was unprepared, it could
happen that a button was released by mistake (A or B), which would make
the message to disappear.

To prevent this, it is now enforced that buttons A and B must be kept
not pressed for at least one frame before it is possible to close the
message box.