Fix graphical glitches when changing minimaps
In the rooms to select scenarios or to generate new maps, when the
player presses a button, the map is cleared before the new one is
loaded.

This wasn't done fast enough and some graphical glitches could be seen
for a frame. This patch makes the code set the palette to just one color
so that the change is done faster than updating the map itself.