Support spaces in the cloned path
Before, the path to the current directory (i.e. the path to which µcity was cloned) would be included as part of each INCLUDE flag. However, this didn’t work if the path to the current directory includes spaces.

To fix this, I’m simply passing `-i$(dir)/` (without the current directory prefix), as it can be found within the current directory without needing to manually add the prefix.