Add compatibility ROM target to Makefile
Some emulators and flashcarts have problems to detect the size of the
SRAM needed by the ROM and either fail to load it or fail to save the
data correctly (resulting in saved data loss).

This patch adds a target to the Makefile to build a compatibility ROM,
which is exactly the same as the regular ROM, but with a more common
(and smaller) SRAM size. This makes it easier for people without the
RGBDS toolchain to get access to a ROM with the fix.