Use ldh [$ff00+c], not ld [$ff00+c], and fix a rRAMG bug
RGBDS 0.9.0 deprecates `ld [$ff00+c]` [1].

Note that this also fixes a bug! One line had `ldh [rRAMG], a` instead
of `ld [rRAMG], a`. Since `rRAMG` is `$0000`, this was quietly getting
assembled as `ldh [$FF00], a` (aka `ldh [rP1], a`) instead of the
desired `ld [$0000], a`. (I think this validates our decision to
deprecate `ldh [$xx]`!)

[1] https://rgbds.gbdev.io/docs/master/rgbasm-old.5#LD__C_,_A_and_LD_A,__C_