Fix DATA_MONEY_AMOUNT macro
When the amount of money is too big it can overflow a 32 bit integer. This
used to cause overflows in old versions of rgbasm. Now, it sets the value
to 0.

This commit fixes the issue by splitting money amounts into two arguments
so that they are always valid 32-bit values.

Thanks @Rangi42 for the bug report!